Add a remote_label to the Vasp evaluation call in `gap_rss_iter_fit.py`.
Now that DFT evaluations are just done using `generic.run`, it's harder to
specify their `RemoteInfo` from an env var with a dict key that points to
the call stack and still distinguish them from other calls to
`generic.run`, because more of the tail end of the call stack is identical.
As a result, it's more useful to have (unique) remote_label arguments
to the egeneric.run` calls that do the DFT evaluations, so add one to
`gap_rss_iter_fit.py`.