Very ugly attempt to fix issues with pickling functions for test_doc_examples.py in github CI.
Part 1: run example code by writing it to a file and loading it as a module with importlib instead of using exec.

Part 2: In the example notebook, add PWD to the PYTHONPATH in each remote_info object or dict, counting on the fact that the test framework will chdir to tmp_path, where it wrote the example code as a file (see Part 1) before the example is run, so PWD is the right place for the extracted code's module.

Pickling works locally, but remote jobs are using the global wfl instead of the one in the directory where pytest is being run from.  The remoteinfo_env fixture deals with this by modifying the PYTHONPATH in the remote_info env_vars field for other remote run tests, but adding that to the example would make it even uglier than these modifications already do.

Should the RemoteInfo constructor itself be able to add such information (e.g. extra bits for the PYTHONPATH env var) from an env var (maybe something like PYTEST_REMOTE_INFO_EXTRAS, or a more PYTHONPATH specific var name), which pytest can then use without polluting the example itself?