Add universal kspacing and refactor handling of nonperiodic cells
Universal kspacing only implemented in Vasp so far.

Refactoring of nonperiodic mainly consists of calculators.generic doing
the removal of stress for fully non-periodic, and cleaning up the DFT
calculators to only do the minimum.

Also, the saving of the results can optionally be done in the
calculator, and Vasp uses that because it changes the pbc, and
that can break the ASE result caching and lead to extra calculations.